Review: simplify the status usage in securityprofilecontext
